+++ src/main/java/uk/gov/justice/artemis/manager/connector/combined/duplicate/BrowsedMessages.java
@@ -1,21 +1,22 @@
 package uk.gov.justice.artemis.manager.connector.combined.duplicate;
 
+import java.util.List;
 import java.util.Map;
 
 import javax.jms.Message;
 
-public class DuplicateMessages {
+public class BrowsedMessages {
 
-    private final Map<String, Message> duplicateMessages;
+    private final Map<String, List<Message>> duplicateMessages;
     private final Map<String, Message> messageCache;
 
-    public DuplicateMessages(final Map<String, Message> duplicateMessages,
-                             final Map<String, Message> messageCache) {
+    public BrowsedMessages(final Map<String, List<Message>> duplicateMessages, final Map<String, Message> messageCache) {
+
         this.duplicateMessages = duplicateMessages;
         this.messageCache = messageCache;
     }
 
-    public Map<String, Message> getDuplicateMessages() {
+    public Map<String, List<Message>> getDuplicateMessages() {
         return duplicateMessages;
     }