+++ src/test/java/uk/gov/justice/artemis/manager/connector/combined/duplicate/TopicDuplicateMessageFinderTest.java
@@ -3,7 +3,6 @@ package uk.gov.justice.artemis.manager.connector.combined.duplicate;
 import static java.util.Arrays.asList;
 import static java.util.Collections.emptyEnumeration;
 import static java.util.Collections.enumeration;
-import static java.util.Collections.singletonList;
 import static java.util.UUID.randomUUID;
 import static org.hamcrest.CoreMatchers.hasItems;
 import static org.hamcrest.CoreMatchers.is;
@@ -14,7 +13,6 @@ import static org.mockito.Mockito.when;
 
 import uk.gov.justice.artemis.manager.connector.combined.CombinedManagementFunctionException;
 
-import java.util.Collection;
 import java.util.Enumeration;
 import java.util.Iterator;
 import java.util.List;
@@ -31,17 +29,16 @@ import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
 
 @RunWith(MockitoJUnitRunner.class)
-public class DuplicateMessageFinderTest {
+public class TopicDuplicateMessageFinderTest {
 
     @Mock
     private JmsMessageUtil jmsMessageUtil;
 
     @InjectMocks
-    private DuplicateMessageFinder duplicateMessageFinder;
+    private TopicDuplicateMessageFinder topicDuplicateMessageFinder;
 
-    @SuppressWarnings("unchecked")
     @Test
-    public void shouldReturnSingleInstanceOfDuplicatedMessages() throws Exception {
+    public void shouldReturnTopicDuplicateMessages() throws Exception {
 
         final QueueBrowser queueBrowser = mock(QueueBrowser.class);
         final Message message_1 = mock(Message.class);
@@ -50,49 +47,7 @@ public class DuplicateMessageFinderTest {
         final Message message_4 = mock(Message.class);
         final Message message_5 = mock(Message.class);
         final Message message_6 = mock(Message.class);
-
-        final String jmsMessageId_1 = randomUUID().toString();
-        final String jmsMessageId_2 = randomUUID().toString();
-        final String jmsMessageId_3 = randomUUID().toString();
-
-        when(jmsMessageUtil.getJmsMessageIdFrom(message_1)).thenReturn(jmsMessageId_1);
-        when(jmsMessageUtil.getConsumerFrom(message_1)).thenReturn("artemis-manager.command_handler");
-        when(jmsMessageUtil.getJmsMessageIdFrom(message_2)).thenReturn(jmsMessageId_1);
-        when(jmsMessageUtil.getConsumerFrom(message_2)).thenReturn("artemis-manager.command_handler");
-        when(jmsMessageUtil.getJmsMessageIdFrom(message_3)).thenReturn(jmsMessageId_1);
-        when(jmsMessageUtil.getConsumerFrom(message_3)).thenReturn("artemis-manager.command_handler");
-        when(jmsMessageUtil.getJmsMessageIdFrom(message_4)).thenReturn(jmsMessageId_2);
-        when(jmsMessageUtil.getConsumerFrom(message_4)).thenReturn("artemis-manager.command_handler");
-        when(jmsMessageUtil.getJmsMessageIdFrom(message_5)).thenReturn(jmsMessageId_3);
-        when(jmsMessageUtil.getConsumerFrom(message_5)).thenReturn("artemis-manager.command_handler");
-        when(jmsMessageUtil.getJmsMessageIdFrom(message_6)).thenReturn(jmsMessageId_3);
-        when(jmsMessageUtil.getConsumerFrom(message_6)).thenReturn("artemis-manager.command_handler");
-
-        final Enumeration<Message> messageEnumeration = enumeration(asList(message_1, message_2, message_3, message_4, message_5, message_6));
-
-        when(queueBrowser.getEnumeration()).thenReturn(messageEnumeration);
-
-        final BrowsedMessages browsedMessages = duplicateMessageFinder.findDuplicateMessages(queueBrowser);
-
-        final Map<String, List<Message>> duplicateMessagesMap = browsedMessages.getDuplicateMessages();
-        assertThat(duplicateMessagesMap.size(), is(2));
-
-        assertThat(browsedMessages.getMessageCache().size(), is(3));
-
-        final Collection<List<Message>> values = duplicateMessagesMap.values();
-        assertThat(values, hasItems(singletonList(message_1), singletonList(message_5)));
-    }
-
-    @Test
-    public void shouldNotReturnTopicDuplicateMessages() throws Exception {
-
-        final QueueBrowser queueBrowser = mock(QueueBrowser.class);
-        final Message message_1 = mock(Message.class);
-        final Message message_2 = mock(Message.class);
-        final Message message_3 = mock(Message.class);
-        final Message message_4 = mock(Message.class);
-        final Message message_5 = mock(Message.class);
-        final Message message_6 = mock(Message.class);
+        final Message message_7 = mock(Message.class);
 
         final String jmsMessageId_1 = randomUUID().toString();
         final String jmsMessageId_2 = randomUUID().toString();
@@ -110,20 +65,24 @@ public class DuplicateMessageFinderTest {
         when(jmsMessageUtil.getConsumerFrom(message_5)).thenReturn("artemis-manager.event_listener");
         when(jmsMessageUtil.getJmsMessageIdFrom(message_6)).thenReturn(jmsMessageId_3);
         when(jmsMessageUtil.getConsumerFrom(message_6)).thenReturn("artemis-manager.event_processor");
+        when(jmsMessageUtil.getJmsMessageIdFrom(message_7)).thenReturn(jmsMessageId_3);
+        when(jmsMessageUtil.getConsumerFrom(message_7)).thenReturn("artemis-manager.event_indexer");
 
-        final Enumeration<Message> messageEnumeration = enumeration(asList(message_1, message_2, message_3, message_4, message_5, message_6));
+        final Enumeration<Message> messageEnumeration = enumeration(asList(message_1, message_2, message_3, message_4, message_5, message_6, message_7));
 
         when(queueBrowser.getEnumeration()).thenReturn(messageEnumeration);
 
-        final BrowsedMessages browsedMessages = duplicateMessageFinder.findDuplicateMessages(queueBrowser);
+        final BrowsedMessages topicBrowsedMessages = topicDuplicateMessageFinder.findTopicDuplicateMessages(queueBrowser);
 
-        final Map<String, List<Message>> duplicateMessagesMap = browsedMessages.getDuplicateMessages();
+        final Map<String, List<Message>> duplicateMessagesMap = topicBrowsedMessages.getDuplicateMessages();
         assertThat(duplicateMessagesMap.size(), is(1));
 
         final Iterator<List<Message>> messageIterator = duplicateMessagesMap.values().iterator();
-        assertThat(messageIterator.next().get(0), is(message_1));
+        final List<Message> messageList = messageIterator.next();
+        assertThat(messageList.size(), is(3));
+        assertThat(messageList, hasItems(message_5, message_6, message_7));
 
-        assertThat(browsedMessages.getMessageCache().size(), is(3));
+        assertThat(topicBrowsedMessages.getMessageCache().size(), is(3));
     }
 
     @Test
@@ -149,7 +108,7 @@ public class DuplicateMessageFinderTest {
 
         when(queueBrowser.getEnumeration()).thenReturn(messageEnumeration);
 
-        final BrowsedMessages browsedMessages = duplicateMessageFinder.findDuplicateMessages(queueBrowser);
+        final BrowsedMessages browsedMessages = topicDuplicateMessageFinder.findTopicDuplicateMessages(queueBrowser);
 
         assertThat(browsedMessages.getDuplicateMessages().isEmpty(), is(true));
         assertThat(browsedMessages.getMessageCache().size(), is(3));
@@ -163,7 +122,7 @@ public class DuplicateMessageFinderTest {
 
         when(queueBrowser.getEnumeration()).thenReturn(messageEnumeration);
 
-        final BrowsedMessages browsedMessages = duplicateMessageFinder.findDuplicateMessages(queueBrowser);
+        final BrowsedMessages browsedMessages = topicDuplicateMessageFinder.findTopicDuplicateMessages(queueBrowser);
 
         assertThat(browsedMessages.getDuplicateMessages().isEmpty(), is(true));
         assertThat(browsedMessages.getMessageCache().isEmpty(), is(true));
@@ -178,7 +137,7 @@ public class DuplicateMessageFinderTest {
         when(queueBrowser.getEnumeration()).thenThrow(jmsException);
 
         try {
-            duplicateMessageFinder.findDuplicateMessages(queueBrowser);
+            topicDuplicateMessageFinder.findTopicDuplicateMessages(queueBrowser);
             fail();
         } catch (final CombinedManagementFunctionException exception) {
             assertThat(exception.getMessage(), is("Failed to browse messages on queue."));